+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# 1.2.1
-No technical change. Add logo and update description in mods.toml
+# 1.3.0
+Add SourcesSelectionScreen to change video and audio source through in-game gui.
+
+Press Ctrl+F12 to open the screen. (the F12 here is definitely the reloadKey in our config so you can customize it too)
+
+Update your Duplicationless to 1.0.13 for compatibility! We are using a new feature in it!